Publish chart feature
Created a publish chart component that uses the dom-to-image library to convert the chart preview element to a blob for saving. The actual saving is done by the library file-saver. The image format is png.

The filename for the saved image is the chart title if specified or defaults to 'published-chart'.

Function toSafeFilename created to convert a string to a safe filename string