Add support for html content above and below the chart svg
To support the rendering of html content above and below the plotly chart we have extended the dynamic chart properties implementation to handle chart properties configured as output: "html". 

These properties will be then be extracted into an htmlProps object that is injected into the chart definition and passed to the chart preview component. The chart preview component can then access htmlProps and render the content into html elements above and below the chart.

Also adjusted plotly margin definition to remove unused space from the title area of the chart