Remove chart title, summary, source and footnotes from chart properties
To align the chart-builder to its future role in the plone-volto CMS the chart title, summary, source and footnotes have been removed from the dynamic chart properties system. 

For the UI/UX demo of chart builder the equivalent text block functionality has been added as simple disconnected components in ChartPreview.tsx. These can be commented out or deleted in future with no functional implications for the rest of chart builder.